ROM can be way bigger than 64k

+++ src/Cart.h
@@ -7,11 +7,11 @@ class Cart {
 public:
   Cart(std::string path);
   void writeToMemory(unsigned char* memory, int bank);
-  long getSize();
+  int getSize();
 
 private:
   unsigned char *data;
-  long size;
+  int size;
   long getFileSize(std::string path);
 };
